Split test suite into multiple sources.  Consolidated some repeated code.  Expanded some tests.  Inlined veque functions.
