+++ performance/main.cpp
@@ -20,7 +20,7 @@ Total:                3,480,344us       10,248,719us        2,492,954us
 
  */
 
-#include "include/veque.hpp"
+#include "veque.hpp"
 #include <algorithm>
 #include <array>
 #include <vector>
